+++ packages/manager/test/manager_observer_test.dart
@@ -1,6 +1,91 @@
+import 'package:manager/src/models/manager.dart';
+import 'package:manager/src/models/observer.dart';
 import 'package:manager/src/models/task.dart';
+import 'package:manager/src/models/task_event.dart';
 import 'package:test/test.dart';
 import 'utils/test_manager_observer.dart' as test_manager_observer;
+
+class _TestGenericAsyncTask<T> extends AsynchronousTask<T> {
+  final Future<T> Function() computation;
+  @override
+  final String id;
+
+  _TestGenericAsyncTask(this.id, this.computation);
+
+  @override
+  Future<T> run() => computation();
+}
+
+class _TestCounterManagerTypeCheckObserver extends ManagerObserver {
+  @override
+  void onCreated(Manager manager) {}
+
+  @override
+  void onDisposed(Manager manager) {}
+
+  @override
+  void onEvent(Manager manager, TaskEvent event) {
+    ManagerObserver.doIfEventIs(event, () => _onDoIfEventCalled++);
+    ManagerObserver.doIfEventIs<TaskSuccessEvent>(
+      event,
+      () => _onDoIfEventCalled++,
+    );
+    ManagerObserver.doIfManagerIs<
+        test_manager_observer.TestCounterManagerTypeCheck2>(manager, () {
+      ManagerObserver.doIfEventIs<TaskLoadingEvent>(event, () {
+        _onDoIfEventCalled++;
+      });
+    });
+    ManagerObserver.doIfEventIs<TaskSuccessEvent>(event, () {
+      ManagerObserver.doIfTaskIs(event.task, () => _onDoTaskIfIsCalled++);
+      ManagerObserver.doIfTaskIs(
+        event.task,
+        () => _onDoTaskIfIsCalled++,
+        whenTaskId: 'one',
+      );
+      ManagerObserver.doIfTaskIs<_TestGenericAsyncTask>(
+        event.task,
+        () => _onDoTaskIfIsCalled++,
+      );
+      ManagerObserver.doIfTaskIs<_TestGenericAsyncTask>(
+        event.task,
+        () => _onDoTaskIfIsCalled++,
+        whenTaskId: 'one',
+      );
+      ManagerObserver.doIfTaskIs<AsynchronousTask<int>>(
+        event.task,
+        () => _onDoTaskIfIsCalled++,
+      );
+      ManagerObserver.doIfTaskIs<AsynchronousTask>(
+        event.task,
+        () => _onDoTaskIfIsCalled++,
+      );
+    });
+  }
+
+  @override
+  void onStateMutated(
+    Manager manager,
+    oldState,
+    newState,
+  ) {
+    ManagerObserver.doIfManagerIs(manager, () {
+      _onDoIfManagerCalled++;
+    });
+    ManagerObserver.doIfManagerIs<
+        test_manager_observer.TestCounterManagerTypeCheck1>(manager, () {
+      _onDoIfManagerCalled++;
+    });
+    ManagerObserver.doIfManagerIs<
+        test_manager_observer.TestCounterManagerTypeCheck2>(manager, () {
+      _onDoIfManagerCalled++;
+    });
+  }
+
+  int _onDoIfManagerCalled = 0;
+  int _onDoIfEventCalled = 0;
+  int _onDoTaskIfIsCalled = 0;
+}
 
 /// It's okay to ignore for test purposes
 // ignore: long-method
@@ -119,5 +204,51 @@ void main() {
       test_manager_observer.TestCounterManager.observer1.hasStateMutated,
       false,
     );
+  });
+
+  group('Testing type check methods of observers - ', () {
+    test(
+        '.doIfManagerIs invokes the callback on sucessful type check of Manager',
+        () {
+      final manager1 = test_manager_observer.TestCounterManagerTypeCheck1(0);
+      final manager2 = test_manager_observer.TestCounterManagerTypeCheck2('0');
+      final observer = _TestCounterManagerTypeCheckObserver();
+      manager1.addObserverTest(observer);
+      manager2.addObserverTest(observer);
+      manager1.run(SynchronousTask.generic(id: 'one', result: 1));
+      manager2.run(SynchronousTask.generic(id: 'one', result: '1'));
+      expect(observer._onDoIfManagerCalled, 4);
+    });
+    test(
+        '.doIfEventIs invokes the callback on sucessful type check of TaskEvent',
+        () async {
+      final manager1 = test_manager_observer.TestCounterManagerTypeCheck1(0);
+      final manager2 = test_manager_observer.TestCounterManagerTypeCheck2('0');
+      final observer = _TestCounterManagerTypeCheckObserver();
+      manager1.addObserverTest(observer);
+      manager2.addObserverTest(observer);
+      manager1.run(SynchronousTask.generic(id: 'one', result: 1));
+      manager2.run(
+        AsynchronousTask.generic(
+          id: 'one',
+          computation: () => Future.value('1'),
+        ),
+      );
+      await manager2.waitForTaskToBeDone(taskId: 'one');
+      expect(observer._onDoIfEventCalled, 6);
+    });
+    test('.doIfTaskIs invokes the callback on sucessful type check of Task',
+        () async {
+      final manager1 = test_manager_observer.TestCounterManagerTypeCheck1(0);
+      final manager2 = test_manager_observer.TestCounterManagerTypeCheck2('0');
+      final observer = _TestCounterManagerTypeCheckObserver();
+      manager1.addObserverTest(observer);
+      manager2.addObserverTest(observer);
+      manager1.run(_TestGenericAsyncTask<int>('one', () => Future.value(1)));
+      manager2
+          .run(_TestGenericAsyncTask<String>('two', () => Future.value('1')));
+      await manager1.waitForTaskToBeDone(taskId: 'one');
+      expect(observer._onDoTaskIfIsCalled, 9);
+    });
   });
 }